Create space-shooter.py
This program sets up a basic space shooter game, with a player's spaceship that can be controlled using the left and right arrow keys, and enemy aliens that move down the screen. The program does not yet include features such as shooting or collision detection, which you can add as needed.

To use this program, you will need to install the pygame library and have an image file named player.png and enemy.png in the same directory as the program. You can customize the game by changing the values of the various game variables, such as the screen size, player and enemy speeds, and image files used.