Update space-shooter.py
In this updated version of the program, the player's spaceship can shoot bullets using the space bar. The bullets move upwards towards the enemy aliens, and when a bullet collides with an alien, both the bullet and the alien are removed from the game.

To use this program, you will need to have the pygame library installed, as well as image files named player.png, bullet.png, and enemy.png in the same directory as the program. You can customize the game by changing the values of the various game variables, such as the screen size, player and bullet speeds, and image files used.